Port external scanner to C
C++ scanners are deprecated according to the offical docs:

    https://tree-sitter.github.io/tree-sitter/creating-parsers
